ensure empty family always remains family_id 0

+++ mcrit/storage/MongoDbStorage.py
@@ -385,11 +385,8 @@ class MongoDbStorage(StorageInterface):
         self._database.samples.delete_one({"sample_id": sample_id})
         # delete family if empty
         family_info = self.getFamily(sample_entry.family_id)
-        if family_info.num_samples == 0:
+        if family_info.num_samples == 0 and family_info.family_id != 0:
             self._database.families.delete_one({"family_id": family_info.family_id})
-            # ensure we always have family_id 0 as empty family.
-            if family_info.family_id == 0:
-                self.addFamily("")
         self._updateDbState()
         return True
 
@@ -445,11 +442,8 @@ class MongoDbStorage(StorageInterface):
             )
             old_family_entry = self.getFamily(sample_entry.family_id)
             # delete family if empty
-            if old_family_entry.num_samples == 0:
+            if old_family_entry.num_samples == 0 and old_family_entry.family_id != 0:
                 self._database.families.delete_one({"family_id": old_family_entry.family_id})
-                # ensure we always have family_id 0 as empty family.
-                if old_family_entry.family_id == 0:
-                    self.addFamily("")
             self._updateDbState()
         if "version" in update_information:
             self._database.samples.update_one({"sample_id": sample_id}, {"$set": {"version": update_information["version"]}})
@@ -474,10 +468,10 @@ class MongoDbStorage(StorageInterface):
             new_num_functions = new_family_info.num_functions + old_family_info.num_functions
             new_num_lib_samples = new_family_info.num_library_samples + old_family_info.num_library_samples
             # update family_entry
-            self._database.families.delete_one({"family_id": family_id})
-            # ensure we always have family_id 0 as empty family.
             if family_id == 0:
-                self.addFamily("")
+                self._database.families.update_one({"family_id": 0}, {"$set": {"num_samples": 0, "num_functions": 0, "num_library_samples": 0}})
+            else:
+                self._database.families.delete_one({"family_id": family_id})
             self._database.families.update_one({"family_id": new_family_id}, {"$set": {"num_samples": new_num_samples, "num_functions": new_num_functions, "num_library_samples": new_num_lib_samples}})
             # update sample_entry and function_entries with new family information
             self._database.samples.update_many({"family_id": family_id}, {"$set": {"family_id": new_family_id, "family": family_name}})
@@ -487,7 +481,6 @@ class MongoDbStorage(StorageInterface):
 
     def deleteFamily(self, family_id: int, keep_samples: Optional[str] = False) -> bool:
         family_document = self._database.families.find_one({"family_id": family_id})
-        family_name = family_document["family_name"]
         if family_document is None:
             return False
         sample_entries = self.getSamplesByFamilyId(family_id)
@@ -497,10 +490,11 @@ class MongoDbStorage(StorageInterface):
         else:
             for sample_entry in sample_entries:
                 self.deleteSample(sample_entry.sample_id)
-        self._database.families.delete_one({"family_id": family_id})
         # ensure we always have family_id 0 as empty family.
         if family_id == 0:
-            self.addFamily("")
+            self._database.families.update_one({"family_id": 0}, {"$set": {"num_samples": 0, "num_functions": 0, "num_library_samples": 0}})
+        else:
+            self._database.families.delete_one({"family_id": family_id})
         self._updateDbState()
         return True
